pyproject.toml
[tool.coverage.run]
branch = true
source = [ "packages/python-app" ]

[tool.coverage.report]
exclude_lines = ['if TYPE_CHECKING:']
show_missing = true

[tool.pytest.ini_options]
addopts = "--cov --cov-report html:'./coverage/./html' --cov-report xml:'./coverage/./coverage.xml' --html='./reports/./unittests/html/index.html' --junitxml='./reports/./unittests/junit.xml'"

[tool.poetry]
name = "astro-python"
version = "1.0.0"
description = "Astro Python"
authors = [ ]
license = 'Proprietary'
readme = 'README.md'

  [[tool.poetry.packages]]
  include = "packages/python-app"

  [tool.poetry.dependencies]
  python = "^3.10.0"
  webui2 = "^2.4.5"
  click = "^8.1.7"
  uvicorn = "^0.27.1"
  asyncio = "^3.4.3"
  aichar = "^1.0.2"
  python-socketio = "^5.11.1"

  [tool.poetry.group.dev.dependencies]
  autopep8 = "2.0.2"
  flake8 = "6.0.0"
  pytest = "7.3.1"
  pytest-sugar = "0.9.7"
  pytest-cov = "4.1.0"
  pytest-html = "3.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
